refactor: move CustomRecipientWarningBanner to CoWSwap
