+++ apps/cowswap-frontend/src/common/pure/CustomRecipientWarningBanner/index.tsx
@@ -1,0 +1,50 @@
+import { Command } from '@cowprotocol/types'
+import { ButtonSecondaryAlt, InlineBanner, InlineBannerProps } from '@cowprotocol/ui'
+
+import styled from 'styled-components/macro'
+
+type CustomRecipientBannerProps = InlineBannerProps & { onDismiss?: Command }
+
+const RecipientBannerContent = styled.div`
+  display: flex;
+  gap: 10px;
+  align-items: center;
+
+  > p {
+    margin: 0;
+  }
+`
+
+export function CustomRecipientWarningBanner({
+  bannerType,
+  borderRadius,
+  orientation,
+  iconSize = 21,
+  iconPadding = '0',
+  padding = '10px 16px',
+  onDismiss,
+}: CustomRecipientBannerProps) {
+  const handleDismiss = () => onDismiss?.()
+
+  return (
+    <InlineBanner
+      borderRadius={borderRadius}
+      orientation={orientation}
+      iconSize={iconSize}
+      iconPadding={iconPadding}
+      bannerType={bannerType}
+      padding={padding}
+    >
+      <RecipientBannerContent>
+        <p>
+          <strong>Caution:</strong> Order recipient address differs from order owner!
+        </p>
+        {onDismiss && (
+          <ButtonSecondaryAlt minHeight={'28px'} onClick={handleDismiss}>
+            Dismiss
+          </ButtonSecondaryAlt>
+        )}
+      </RecipientBannerContent>
+    </InlineBanner>
+  )
+}
