+++ libs/ui/src/containers/InlineBanner/banners.tsx
@@ -1,13 +1,9 @@
-import { Command } from '@cowprotocol/types'
 import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 
-import styled from 'styled-components/macro'
-
-import { ButtonSecondaryAlt } from '../../pure/ButtonSecondaryAlt'
 import { TokenAmount } from '../../pure/TokenAmount'
 import { Nullish } from '../../types'
 
-import { InlineBanner, InlineBannerProps } from './index'
+import { InlineBanner } from './index'
 
 export enum BannerOrientation {
   Horizontal = 'horizontal',
@@ -34,52 +30,6 @@ export function SmallVolumeWarningBanner({ feePercentage, feeAmount }: SmallVolu
       {/*<br />*/}
       {/* TODO: add link to somewhere */}
       {/*<a href="/">Learn more ↗</a>*/}
-    </InlineBanner>
-  )
-}
-
-type CustomRecipientBannerProps = InlineBannerProps & { onDismiss?: Command }
-
-const RecipientBannerContent = styled.div`
-  display: flex;
-  gap: 10px;
-  align-items: center;
-
-  > p {
-    margin: 0;
-  }
-`
-
-export function CustomRecipientWarningBanner({
-  bannerType,
-  borderRadius,
-  orientation,
-  iconSize = 21,
-  iconPadding = '0',
-  padding = '10px 16px',
-  onDismiss,
-}: CustomRecipientBannerProps) {
-  const handleDismiss = () => onDismiss?.()
-
-  return (
-    <InlineBanner
-      borderRadius={borderRadius}
-      orientation={orientation}
-      iconSize={iconSize}
-      iconPadding={iconPadding}
-      bannerType={bannerType}
-      padding={padding}
-    >
-      <RecipientBannerContent>
-        <p>
-          <strong>Caution:</strong> Order recipient address differs from order owner!
-        </p>
-        {onDismiss && (
-          <ButtonSecondaryAlt minHeight={'28px'} onClick={handleDismiss}>
-            Dismiss
-          </ButtonSecondaryAlt>
-        )}
-      </RecipientBannerContent>
     </InlineBanner>
   )
 }
